+++ riscv/CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Use CSR helper macros to define `mideleg` register
 - Use CSR helper macros to define `mcounteren` register
 - Use CSR helper macros to define `mie` register
+- Use CSR helper macros to define `mimpid` register
 
 ## [v0.12.1] - 2024-10-20